Two-level stack state deserialization
When deserializing stack state from stack.json blobs, a RawStackState is
first constructed. The RawStackState contains Oids which are ensured to
be syntactically correct, but do not necessarily resolve to objects in
the repository.

The RawStackState is then converted into the final StackState by
resolving Oids to Objects in the context of an active Repository.

Resolving Oids to Objects in StackState helps surface state errors
earlier.