Test for generated patchname with trailing period
There is a corner case where generated patchnames can end up with a
trailing period, which is invalid for patch (and git ref) names.

When constructing the shortened name, we now remove trailing periods
from each "word" component. A message such as "This is. My patch." will
now become "this-is-my-patch" instead of "this-is.-my-patch".

Signed-off-by: Peter Grayson <pete@jpgrayson.net>