Tests redirect to err file when grepping stderr
The old pattern:

    command_error stg xxx 2>&1 | grep -e "..."

had the problem that the command_error testing would not abort the test
if stg returned the wrong code. The command_error error message would
print out, but the test would merrily continue as long as grep
succeeded.

The new pattern is:

    command_error stg xxx 2>err && grep -e "..." err

which properly separates the command_error test from the grep test and
has the debugability benefit of being able to insert test_pause between
the stg command and the grep, and being able to inspect err at that
point without having to re-run the stg command.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>