Use --diff opt in `stg new`
Compat: having --edit and --diff for `stg new` is a new feature. But
--diff sort of conflicts with --verbose, which was used to include the
diff in the interactive editor, but only if the user did not provide a
message on the command line. I.e. there used to not be a way to both
provide a message on the command line *and* invoke the editor.