Normalized patch edit command line options
The new patchedit module pulls together all the various patch editing
command line argument definitions along with an edit() function that can
interpret and execute on those arguments.

The end-game is for new, refresh, squash, and edit commands to all have
all of these editing options in common. As things stand in the Python
implementation, there is a lot of overlap among these commands, but with
notable and arbitrary differences.