Punt to python with python instead of pyo3
The original intent was to use pyo3 to reach into StGit's python
implementation internals to bridge gaps between the rust and python
implementations, but insofar as the pattern is to punt to python in a
wholesale fashion on a command-by-command basis, that can be trivially
accomplished by invoking a python interpreter process. Easy. And drops a
bunch of dependencies.