Preserve picked patch name when possible
When `stg pick` picks a patch, it now preserves the patch name as long
as it is unique to the stack it is being brought into. Previously, patch
names with certain valid, but non-word, non-digit characters (e.g. '@')
would be morphed such that the original patch name was not preserved.

Repairs #175

Signed-off-by: Peter Grayson <pete@jpgrayson.net>