Improve patch name generation
There were some cases where StGit would generate invalid patch names for
new and/or imported patches. Notably:

   stg new -m "a:.b.:"

would fail due to the generated patch name having a trailing '.'.

The regex based approach to transforming a raw message or filename into
a patchname is replaced with an approach that transforms the raw string
on a character by character basis.

Repairs #176

Signed-off-by: Peter Grayson <pete@jpgrayson.net>