Use anyhow for errors
There turn out to be relatively few control flow decisions made on errors.
There are also a lot of single-use errors/error-strings. Combined, this
means that definig custom errors for everything does not net a lot of
value.

Application errors is anyhow's sweet spot, and so it seems to fit well with
StGit's needs.

Some custom errors are retained in the stack::error module.