Broken pipe error message for `stg log`
When `stg log`'s underlying `git log` process uses a pager, and that pager
exits before `git log` can write all its output, the `stg log` process
exits with an empty error message because `git log` does not print an error
for this case.

Detect when the subordinate `git log` process exits due to a SIGPIPE signal
and indicate this reason in the user-facing error message.