Distinguish editable and edited patch descriptions
Instead of using a single PatchDesription structure to represent both the
outgoing/editable patch description and the incoming/user-edited
description, now use two different structures.

The specific motivation is to be able to distinguish the cases from the
user-edited patch description:

- "Patch:" header present with patch name value
- "Patch:" header present with empty patch name value
- "Patch:" header absent

When the "Patch:" header is absent, the patch name now reverts to the
original patch name, whereas when the "Patch:" header is present with an
empty value the patch name is generated from the message.

The driving assumption is that the "Patch:" header being entirely absent
indicates that the user is just writing a message and isn't trying to
rename or otherwise alter the patch name.

On the other hand, when "Patch: " is found with an empty value, the user
explicitly deleted the pre-filled patch name. This is interpreted as the
user desiring for the patch name to be regenerated from the message. If the
user had wanted to retain the original patch name, they would have simply
left the "Patch:" header and value unaltered.