Remove `stg mail`
This is the last command that relied on the Python implementation of StGit,
but it is not being implemented in Rust, at least for the time being.

The rationale is that it is not clear that `stg mail` provides sufficient
value versus running `git send-email`. Consider, for example:

    $ stg mail p0..p4

The alternative using git would be:

    $ git send-email $(stg id p0)..$(stg id p4)

Or:

    $ git send-email refs/patches/main/p0..refs/patches/main/p4

N.B. `git send-email` provides a very wide selection of options, both on
the command line and in the config, to fine-tune how emails are sent. `stg
mail` only offered a subset of those capabilities.

Perhaps the most obvious feature that `stg mail` had that `git send-email`
does not [quite] have is customizable patch templates. This feature,
however, may be more trouble than its worth since it allows patch emails to
deviate from git email standards. Note that git allows the branch
description to be used as the cover letter (i.e. with
--cover-from-description), so that seems like a viable replacement for `stg
mail`'s cover letter template.