Add patchrange::parse_single()
This function parses a single patch name argument, verifying that it is
known to the stack and is in an allowed list (applied, unapplied, hidden).
It also performs string similarity matching on unknown patches and provides
user-facing error messages with the similar patch names.

This isolates use of the strsim crate to the patchrange module.