Zsh command and alias completions from stg -h
The Zsh completions script is updated such that StGit subcommands and
aliases are now derived from the `stg -h` output from the Rust
implementation.

The stg executable is now fast enough that caching the parsed command list
is no longer necessary.

And since the description of aliases contain the complete alias expansion,
the completion script can also derive what it needs to know about aliases
from the `stg -h` output without having to consult `git config`. This works
for both StGit's built-in aliases as well as user-defined aliases, which
means the completions script no longer needs baked-in knowledge of the
built-in aliases.