Add copyright and license to `stg version`
The --version option is now handled by StGit instead of delegating to clap.
The -V option, which was not present in the Python version, is eliminated.

A new -s/--short option is added to `stg version` to print abbreviated
version information (i.e. without copyright and license notices or git
version info). The Zsh completions are updated accordingly.

And clap's "cargo" feature is no longer used, so Cargo.toml is updated to
reflect that.