Implement `stg completion bash`
This dynamically generates a bash completion script.

In order to complete StGit specific objects such as patches and branches,
the script generation code needs to somehow know arguments' (both
positionals and options) value kinds. Although clap allows a ValueHint to
be associated with an Arg, ValueHints can only specify generic value kinds
and not application-specific kinds. The solution being used is to make the
Arg id (i.e. the string provided to Arg::new()) meaningful. Note that the
string provided to Arg::value_name() is different and is *not* used when
generating completions. See insert_compreply() for this mapping in action.

N.B. The completion code for the top-level `stg` command and its options
are hard-coded. This is needed for special handling of the -C option as
well as aliases. Thus any changes to top-level and/or global options will
require manual changes in the bash completion generator.