Zsh more/better tests for stg/git command success
These tests are important so that the user gets appropriate feedback when
attempting completions outside the context of a git repository and/or an
initialized StGit branch.

N.B. __stg_command_successful is used when running stg commands that
require an initialized stack whereas __stg_git_command_successful is used
for all other git and stg commands.