Zsh better patch completions
Patch completions now use `stg series` output as description strings. The
user thus now sees the applied/top/unapplied/hidden sigil ('+', '>', '-',
'!') in addition to the patch name in the offered completion list. And the
order of the completions is now in series order instead of allowing zsh to
sort the patch names lexically.

The completions are also forced to be in a single-column list. The patch
completion list looks an awful lot like the output of `stg series` now.

- Completions for `stg series` is repaired to allow multiple patches to be
  selected.
- Completions for `stg show` are cleaned-up as well.