Do not use `git apply --allow-empty`
The --allow-empty option to `git apply` is relatively new, first appearing
in git 2.35.0. Although its behavior is pretty ideal for many StGit
applications, it needs to be avoided to maintain compatibility with (only
slightly) older git versions.

Also repair the related issue where the failure of the `git apply` command
was misattributed to `git diff-tree`. This could occur in cases where the
output of a diff-tree process was routed to the stdin of an apply process.
The solution is to explicitly test the return status of the downstream
apply process before testing the status of the upstream diff-tree process.

Repairs #192