Generate asciidoc man pages
Man pages are generated as asciidoc in much the same manner and format as
the Python implementation.

A key difference/improvement versus the Python implementation is that StGit
commands with their own subcommands are specifically handled. When a
command has subcommands, a "COMMANDS" section is added to the man page and
the OPTIONS section has separate subsections for each subcommands options.

Another difference is that "linkstg" and "linkgit" asciidoc macros
automatically replace certain forms found in commands' about strings and/or
their options' help strings. For example, a single-quoted string `'stg
series'` will be converted to `linkstg:series[]`. When `asciidoc` or
`asciidoctor` is used to process the asciidoc files, these link macros are
further converted to a form appropriate for the final output format. E.g.
these are turned into hyperlinks when the output format is html.

The last key difference from the Python implementation is that asciidoc
files for all StGit commands are generated in a batch. This avoids the
overhead using `cargo run` for each StGit subcommand. In order to be more
make-friendly, existing asciidoc output files are only overwritten if the
content changes, thus preserving timestamps for any unchanged files. The
internal asciidoc generation is sufficiently fast that the cost of running
`cargo run` dwarfs both the generation time and the time to compare with
existing files.