Use STG_PROFILE make variable everywhere
The STG_TEST_PROFILE variable is replaced with STG_PROFILE in t/Makefile
and STG_PROFILE is introduced into the other Makefiles.

This solves the problem of only needing to build either the dev or the
release profile, but not both when invoking top-level install targets.
Without explicitly specifying the profile to various cargo commands, some
would use the dev profile by default (e.g. `cargo build` and `cargo run`)
while others (`cargo install`) default to the release profile.

The top-level Makefile defaults to using the release profile. t/Makefile
falls back to the dev profile, but inherits from the top-level when invoked
recursively. This means that when cd'ed into t/, the dev build will be
used, which is probably what you want if you're cd'ed into t/.