Improved error when push conflicts with untracked
When pushing a patch that adds a file that conflicts with an untracked file
in the working tree, the error stated "Index/worktree dirty" without
further elaboration.

The error message from `git merge-recursive` is now displayed. This message
indicates the particular untracked files that prevent the push's merge from
completing.

Also add test cases to ensure this behavior.

Addresses #193