Repair spilling added files with path limits
When using `stg spill` with a patch containing newly added files, those
newly added files were not spilled when a pathlimit was present on the
command line.

This issue is repaired by using a different and more robust strategy for
creating the updated (spilled) index for the patch.

Relevant test cases are added to t2705-spill.sh.