Show git commit id with version when needed
The version of StGit displayed by `stg version` is primarily informed by
the package version from Cargo.toml. A problem, however, is that only for
proper (tagged) StGit releases will the package version correspond with the
git tag from which it was built. Thus any one-off build of StGit will lie
about its version.

To solve this, a build.rs script is added that looks at `git describe`
output to determine whether the `stg version` version should be
supplemented with the git commit id (hash) from whence it was built.

When StGit is built from a tag and that tag matches the package version
from Cargo.toml, the git hash is elided from `stg version`. Otherwise, `stg
version` shows the hash as a parenthetical supplement to the
Cargo.toml-supplied version. If the working tree is dirty, the hash is
supplemented with an asterisk.

When building from outside a git repo, e.g. from a release .tar.gz archive,
the build script falls back to using the only version available to it: the
package version from Cargo.toml.