Repair stgit.el patch diff
In StGit 2.0, `stg show` takes optional pathspec arguments after a `--`
separator. In earlier versions of StGit, the patch name arguments could
come either before or after the [optional] `--` separator.

stgit.el was invoking `stg show` with the patch name argument after `--`.
This was not necessary for older versions of StGit and does not work with
StGit 2.0.

The issue is resolved by omitting the `--` separator when composing the
`stg show` command line.

Fixes: #202