Ensure rerun-if-changed files exist in build.rs
When using the "cargo:rerun-if-changed=<path>" directive in the build.rs script,
the absense of the path triggers cargo to rebuild.

The .git/packed-refs file does not always exist in a git repository;
especially a fresh repository. This was causing CI builds to take a very
long time due to cargo having to relink the stg executable many, many times
for `make doc`.

This problem is solved by having the build script test the existance of
each "rerun-if-changed" file candidate.