Compile-time builtin commands list
The list of builtin StGit commands was being generated as a BTreeMap at
runtime, but since all the information contained in that BTreeMap is known
statically at compile time, it was possible to reexpress the list of
builtins as a compile time generated slice.

This approach is generally simpler, with less code and possibly less
runtime overhead. One caveat is that the commands are now searched linearly
(O(n)) instead of the BTreeMap which has O(ln(n)) searches, but command
name search is typically only performed once per run and the total number
of commands is small enough that a materially deleterious effect on runtime
performance is not expected.