Upgrade to git2 0.15
Since 0.13, a few changes have been made to git2, including
https://github.com/rust-lang/git2-rs/pull/791 which should address the
extensions.worktreeConfig issue in #195.

The upgrade required fixing uses of [ConfigEntries][1] because it's no
longer an iterator.

[1]: https://docs.rs/git2/0.15.0/git2/struct.ConfigEntries.html

However, it provides an API that lends itself well to `while let`
to provide a for loop like experience, so this was an easy fix.