Refactor transaction checkout and execute
The goal for this refactor is for execute() to do its business with the
decomposed parts of a StackTransaction. The reason we want execute() to
operate on the transaction parts instead of the whole struct is that it
enables rearrangement of the order of the various execution steps. And the
reason to rearrange the various execution steps is because there are
failure paths that currently lead to inconsistent stack and repo state that
need more robust rollback.

This refactor prepares the pitch, but does not change behavior.