Transaction rollbacks for SIGINT
When the user sends a SIGINT signal (i.e. by pressing ctrl-c) while the
stack transaction is executing, the signal is now deferred such that the
rollback can occur as it does for other transaction execute time errors.
SIGINT signals outside of the execute-time critical section are not
deferred and will cause the stg process to exit immediately.

The implementation uses the ctrlc crate, which is cross platform and should
also work on Windows.

Also add test case that simulates the stg process being hit with a SIGINT
at transaction execute time, which is a thing that can happen depending on
the gpg pinentry configuration and if the user hits ctrl-c.

N.B. this change only handles SIGINT. The more violent SIGKILL will
immediately terminate the stg process, bypassing the rollback mechanism.

Fixes: #205