Repair gpg test that kills grandparent stg process
This test case is a bit tricky. Most of the time, our custom gpg stand-in
script successfully kills its grandparent stg process with a SIGTERM signal
during the stack transaction execute critical section. However, every once
in a while something else happens where the stg process receives the
SIGTERM outside of its critical section. In this alternate case, the test
is allowed to pass if the repository consistency checks pass.