Allow \- escape for patch names with leading hyphen
Patch names are allowed to start with a hyphen ('-'), but there are several
commands (e.g. `stg new` and `stg show`) where it is not possible to
specify such a patch name on the stg command line because the leading dash
is ambiguous with a command line option and thus is flagged as a command
line parsing error.

Many stg commands allow patch names to come after a "--" separator. But
commands that take both patch names and pathspecs require the pathspecs to
come after the "--" separator.

This problem is solved by allowing patch names and patch ranges that start
with a hyphen to be specified on the command line by escaping the leading
hyphen with a single backslash. This is safe and unambiguous because the
backslash is forbidden from patch names (and Git reference names) and thus
when used in a "\-" prefix can only have the valid interpretation of being
an escaped hyphen.

N.B. in most shells, the backslash itself needs to be escaped with a
backslash or enclosed in single quotes. E.g.:

    stg new \\-patch

or...

    stg new '\-patch'

N.B. patch names with a leading hypen appearing after a "--" separator are
not required to have a backslash-escaped hyphen.