Add --patch option to `stg show`
This option provides an alternative way to express patch name and/or patch
ranges to `stg show`.

The --patch option allows any valid patch name or range to be provided as
its value whereas the patch-or-rev option requires patch names starting
with '-' to be escaped.

Scripts and editor extensions should prefer using the --patch option
instead of the patch-or-rev argument because it obviates the need to check
and possibly escape patch names being provided to `stg show`.

Relates-to: #216