fix(zsh): Repair broken completion of --git-opt
The brace expansion for iterating over the `--git-opt` values from the
command line (i.e. `{2..2..$(($#git_opts - 2))}`) was broken in a couple
ways.

First, in `{n1..n2..n3}`, n2 is the end number and n3 is the step value,
but these were transposed.

Second, in order to avoid backwards sequences, we need to ensure that
the end number is greater than the start number. A condition is added to
test that.