fix(email): Send using --branch option
The refspec passed to `git send-email` would use "HEAD" in several
circumstances instead of an explicit commit id. This caused problems
when the --branch option was used since HEAD refers to the current
branch and not the user-specified branch.

The repair is to use explicit commit ids in all circumstances.

Refs: #182