test: Use prereq test for git send-email
Some platforms (e.g. Ubuntu) do not package `git send-email` with the
rest of git. And some CI platforms (e.g. GitHub Actions) use a version
of git that is incompatible with the distro-provided git-email package.

So since we cannot count on the availability of `git send-email` in all
environments, add GITSENDEMAIL prereq checks to relevant tests that need
`git send-email`.