fix: Avoid duplicate signoff with stgit.autosign
With stgit.autosign configured, duplicate Signed-off-by trailers would
added when refreshing a patch. In the Python implementation, the
stgit.autosign configuration only affected `stg new` and `stg import`, so
the fix is to use that same policy.

N.B. also considered was to use the --if-exists=addIfDifferent option with
`git interpret-trailers`. This option would work to avoid duplicate
trailers, but may still cause trailers to be added unexpectedly when
running commands such as `stg refresh`, `stg edit`, `stg pick`, etc.

Fixes: #227