feat: Stupid version tests
Introduce a mechanism to gather, cache, and test the git executable version
when running subordinate git commands. This used to determine whether the
--3way option to `git apply` may be used in conjunction with `--cached`.

Getting the git version is relatively inexpensive compared to other git
commands, but does involve the overhead of running a subordinate process.
This overhead is mitigated by caching the git executable version in the
StupidContext. The runtime performance impact from this approach appears
minimal.

Fixes: #225