chore: Update for chrono deprecations
The non-deprecated east_opt() and timestamp_opt() interfaces return
Option/Result, but these changes stick with the expect/unwrap approach to
those outcomes. I.e. this change maintains StGit's status quo which is to
panic if a timestamp found in the git repo has an out-of-bounds timezone or
unix timestamp.