fix: StGit-specific branch config handling
When renaming, deleting, or cleaning-up branches, the branch and StGit
specific configuration section, i.e. "branch.<branch>.stgit" was not being
handled correctly.

For `stg branch --rename`, this config section was not being renamed along
with the branch.

For `stg branch --delete` and `stg branch --cleanup`, this section was not
being entirely removed; the section heading would remain, but with no
values.