fix: Don't generate new patch name until after edit
When a patch:

1. Is being being edited interactively
2. And does not yet have a name
3. And a message is not provided on the command line

Do not pre-generate a patch name before the interactive edit session. This
leaves the 'Patch:' header blank for the interactive edit, which allows the
user the flexibility to either fill-in their own patch name or let the
patch name be generated from the message *after* the interactive edit
session.

Fixes: #239