fix: Use GIT_PREFIX in built-in aliases
The built-in aliases 'add', 'mv', 'resolved' and 'rm' all historically,
prior to 438fb8efc, operated from the current working directory and not, as
it is without this change, from the top of the work tree. These aliases are
modified to run git in the $GIT_PREFIX directory to restore the more
natural and historic behavior.