refactor: StackAccess trait
Add new StackAccess trait to complement StackStateAccess. The new
StackAccess trait is a supertrait of StackStateAccess. In addition to the
behaviors provided by the existing StackStateAccess, the new trait provides
accessor methods related to branch state.

This fixes the problem where the StackStateAccess::base() impl for
StackState would panic if ever called. The base() method is now moved to
StackAccess, which is not implemented by StackState, thus giving a strong
compile-time avoidance of this panic.