fix: Accept full ref name for branches
Previously only short branch names such as "master" or "main" could be used
successfully with --branch and --ref-branch options. With this change,
fully qualified reference names such as "refs/heads/master" or
"refs/heads/main" may also be supplied.

N.B. a new error message is introduced for the case where a valid reference
name is supplied, but that reference is not actually a branch.