fix(prev): Different error message for empty stack
When running `stg prev` on an empty stack, the error message will now be
"No patches applied". This new message is more specific and thus offers a
bit more information to the user about the stack's state.

The "Not enough patches applied" error message will still be be shown when
`stg prev` is called with exactly one applied patch.