fix: Improved error message for uninitialized stack
The previous error message of "Branch `<name>` not initialized" is not
specific about what is not initialized and could easily be misconstrued by
a user to mean that the Git branch itself is in a bad/uninitialized state.

The new error message attempts to clarify that it is the StGit stack that
is not initialized for the branch.