feat: Add --committer-date-is-author-date option
Without this new option, any operation that makes a change to a patch's
commit will cause the patch's new commit to have a committer date of the
current date/time. Using --committer-date-is-author-date causes the
committer date to be the same as the author date for any patch commits
created during the command.

N.B. the committer date will neither be updated to the current date or set
to the author date if the command does not have create or update a patch's
commit. E.g. pushing patches that do not require merges does not cause any
new commits to be made and thus the committer date will not be changed
regardless of whether the --committer-date-is-author-date is used.

This option name comes from `git rebase`.

Fixes: #47