fix: Allow "---" separator in messages
A scissors comment line, "# ---- >8 ----", is used instead of the "---"
separator to separate the patch headers and message from the diff in a
patch description buffer. This allows "---" to be in the regular commit
message body.

This was a regression from StGit 1.x. But StGit 1.x also had a problem
with "---" separators when using `stg edit -d`. The cut line comment
approach allows `stg edit -d` to work as well.

Fixes: #243