test: Workaround bug in git 2.39.0
git 2.39.0 has a weird little bug where unmerged files in the output of
`git diff --cached --stat` are missing newlines. This breaks a particular
test. The fix is to limit the diff output to the particular file of
interest so that the missing newline on other files in the stat output
don't break the grep.