refactor: Use gitoxide instead of git2
gitoxide's git_repository crate provides similar capabilities to the git2
crate. The primary advantage of gitoxide is that it is faster than git2;
specifically for startup time where it is about 4x faster, and possibly for
other operations.

gitoxide is also written in pure Rust, which provides a richer set of
interfaces and opens the door to more direct participation in upstream
improvements. Also, gitoxide aims to implement more git behaviors than
libgit2, and thus provides a path to eliminate calls to git which would
further improve runtime performance of StGit.

The potential downside to gitoxide is that it introduces a significant
numbmer (40) of additional Rust crate dependencies.