fix: Error if author or committer is not configured
Although git has a mechanism to guess author and committer identities, its
porcelain commands, such as `git commit`, require the user to be
configured. Follow suit in StGit by raising an error in this circumstance
instead of using gitoxide's fallback identity. This is also more consistent
with StGit's pre-gitoxide behavior.

Relates-to: #256