feat(series): override display options
Add new "--no-xxx" options for the various options in the "Display Options"
group. This allows any of the display options to be turned on or off by
suffixing the command line with one of these options.

This may be especially useful for aliases of `stg series`. For example an
alias of "listc" mapping to "series --commit-id" could then be called with
"stg listc --no-commit-id" to disable the display option baked into the
alias. Without these "--no-xxx" options, that would not be possible.