fix(rename): colliding patch names
A colliding patch name is one that differs only by character case from
another patch in the stack.

When renaming a patch, the new name must not collide with other names in
the *updated* stack. The problem was that any collision in the current
stack was being flagged as an error, but the new name colliding with the
patch being renamed should be allowed. Thus renaming, for example, "foo"
to "Foo" would fail.

This case is repaired for `stg rename`. Furthermore, the
StackTransaction::rename_patch() method is improved to cover the same issue
to cover renames from `stg edit` and `stg rebase -i`.