fix: run commit-msg hook in work dir
The commit-msg hook is supposed to be run from the work tree root when not
operating on a bare repository.

Instead of using tempfile::NamedTempFile, we switch to creating our own
temporary file relative to the work dir. This allows the temporary file
argument to commit-msg to be a simple relative file name, which eases the
complexity of passing the argument on Windows.