fix(import): lost subject lines resembling header
Using `git mailinfo` when importing patch files (as opposed to mail/mbox
files) often works well, but there is at least one case that leads to a
regression vs. StGit 1.x. Specifically, if the patch file contains a
subject line of the form "something: more" (i.e. contains a ':'), `git
mailinfo` interprets it as an unknown header and discards it. Thus the
subject line does not show up in any of the header info, message, or diff
outputs of `git mailinfo`.

The approach taken is to use more custom patch parsing as was done in StGit
1.x. Thus `git mailinfo` is no longer used for regular patch file imports,
but is still used when --mail or --mbox is provided.

Fixes: #321