fix(branch): create based on remote branch
Creating a branch with the base committish being a remote branch (e.g.
origin/master) was not handled correctly and would trigger an assert.

With this change, remote base branches are handled without asserting. The
behavior, however, is somewhat different than that of StGit 1.x with
respect to how the branch.<name>.remote and branch.<name>.merge
configuration is setup. In StGit 1.x, when using a remote base branch,
those keys were not set. With this change, branch.<name>.remote is set to
the remote part of the provided base branch and branch.<name>.merge is set
to the remainder, i.e. given "origin/master" branch.<name>.remote would be
"origin" and branch.<name>.merge would be "master".

The idea is that when the user provides a remote base branch, their intent
is for the created branch to track that remote branch.

Fixes: #317