fix!: uncommit checks for HEAD/top mismatch
Relevant historical commits: bca31c2 and 9beaff2.

There was an issue where `stg uncommit` would reset HEAD to the stack top
when there was a HEAD/top mismatch; i.e. where a `git` command causes HEAD
and the stack top to diverge such that a follow-up `stg uncommit` would
then reset HEAD to the old/wrong stack top.

The solution at the time was to have `stg uncommit` ignore the HEAD/top
mismatch, go ahead with uncommitting patches, and explicitly *not* set the
branch HEAD. This works out because uncommit intrinsically only adds
patches to the bottom of the stack, so whatever is going on at the top/HEAD
can be viewed as "not it's problem".

But just because `stg uncommit` *can* ignore an existing HEAD/top mismatch,
doesn't mean it should. Which is the core of this issue.

The solution is to add the HEAD/top mismatch check (which is also used by
other commands that mutate the stack) so that user will know there is a
mismatch that needs repair before the uncommit is allowed. We will do that
check even though the `stg uncommit` operation *can* perform its nominal
behavior even when a HEAD/top mismatch exists.

All of the above is elaborated in recognition that this change has the
potential to break workflows, but we nonetheless choose to make `stg
uncommit` more restrictive due to the belief that detecting HEAD/top
mismatch earlier is going to be, in net, more helful to StGit users.

Fixes: #360