fix: improved interactive editor diagnostics
Add context to spawning and waiting for the child interactive editor
process. This will make it more apparent which step of the process
failed when there is a failure.

Also ensure that the "waiting for your editor" hint is cleared before
allowing any error to be shown to the user. This ensures that the error
output appears on its own line in the user's terminal.